requirements.txt
python-telegram-bot
stellar-sdk
